requirements.txt
python-telegram-bot==13.15
pandas==2.1.4
numpy==1.26.0
python-dotenv
urllib3<2.0
six
